chore: rename class
