+++ app/src/main/java/com/example/judgement/view/main/scrap/ScrapManager.kt
@@ -1,7 +1,7 @@
 package com.example.judgement.view.main.scrap
 
-import com.example.judgement.adapter.ScrapRvAdapter
+import com.example.judgement.adapter.ScrapAdapter
 
 data class ScrapManager(
-    val scrapRvAdapters: MutableMap<String, ScrapRvAdapter> = mutableMapOf(), // 어댑터
+    val scrapRvAdapters: MutableMap<String, ScrapAdapter> = mutableMapOf(), // 어댑터
 )
